+++ wishing-well/src/main/java/win/amos698/sweettools/wishingwell/repo/WishRepo.java
@@ -6,12 +6,12 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import win.amos698.sweettools.wishingwell.repo.model.Wish;
 
-public interface WishRepo extends JpaRepository<Wish, Integer> {
+public interface WishRepo extends JpaRepository<Wish, Long> {
 
     @Query(value = "SELECT * FROM (SELECT * FROM wish WHERE userId=:userId) ORDER BY RAND() LIMIT 1", nativeQuery = true)
-    Wish pickOne(@Param("userId") int userId);
+    Wish pickOne(@Param("userId") long userId);
 
     @Modifying
     @Query(value = "UPDATE WISH SET used=0 WHERE wishId=:wishId", nativeQuery = true)
-    void deleteWish(@Param("wishId") int wishId);
+    void deleteWish(@Param("wishId") long wishId);
 }
